refactor: import changes and store creating
