+++ src/app/pages/movie-now-top-rate-page/movie-now-top-rate-page.component.ts
@@ -5,7 +5,7 @@ import { MovieCardComponent } from '../../components/movie-card/movie-card.compo
 import { Movie } from '../../models/movie.model';
 import { Subscription } from 'rxjs';
 import { Store } from '@ngrx/store';
-import { selectTopRatedMovies } from '../../store/selectors';
+import { selectTopRatedMovies } from '../../store/movie-store/selectors';
 
 @Component({
   selector: 'app-movie-now-top-rate-page',
